Attempt to disable log collection in task driver
This commit attempts to disable log collection in a custom Nomad task
driver based on https://github.com/hashicorp/nomad-skeleton-driver-plugin.

The job created is not running (and will fail), but the expected result
is Nomad not starting a `nomad logmon` process every time the job is
restarted — which is still happening.

This is a work in progress.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>